+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Replaced rome datepicker with flatpickr as rome is no longer maintained, as a result moment is no longer available through the designguide.
 - Changed cards so it is no longer required to wrap cards in the `.cards` container for them to work.
+- Text color utility class `.color-{name}` has been replaced with `.text-{name}`, and is deprecated.
 
 ### Fixed